Conflicts:
	src/red.cuda.initial/main.cu

+++ src/red.cuda.benchmark/main.cu
@@ -1015,10 +1015,10 @@ void cpu_calc_grav_accel_naive(int n_body, const var4_t* r, const var_t* mass, v
 
 void cpu_calc_grav_accel_naive(interaction_bound int_bound, const var4_t* r, const var_t* mass, var4_t* a)
 {
-	for (int i = int_bound.sink.x; i < int_bound.sink.y; i++)
+	for (unsigned int i = int_bound.sink.x; i < int_bound.sink.y; i++)
 	{
 		var4_t dVec = {0.0, 0.0, 0.0, 0.0};
-		for (int j = int_bound.source.x; j < int_bound.source.y; j++) 
+		for (unsigned int j = int_bound.source.x; j < int_bound.source.y; j++) 
 		{
 			if (i == j)
 			{
@@ -1083,10 +1083,10 @@ void cpu_calc_grav_accel_naive_sym(int n_body, const var4_t* r, const var_t* mas
 
 void cpu_calc_grav_accel_naive_sym(interaction_bound int_bound, const var4_t* r, const var_t* mass, var4_t* a)
 {
-	for (int i = int_bound.sink.x; i < int_bound.sink.y; i++)
+	for (unsigned int i = int_bound.sink.x; i < int_bound.sink.y; i++)
 	{
 		var4_t dVec = {0.0, 0.0, 0.0, 0.0};
-		for (int j = i + 1; j < int_bound.source.y; j++) 
+		for (unsigned int j = i + 1; j < int_bound.source.y; j++) 
 		{
 			// r_ij 3 FLOP
 			dVec.x = r[j].x - r[i].x;
@@ -1122,10 +1122,10 @@ void cpu_calc_grav_accel_naive_sym_advanced(interaction_bound int_bound, const v
 	const int n_source = int_bound.source.y - int_bound.source.x;
 	const int n_body = min(n_sink, n_source);
 
-	for (int i = int_bound.sink.x; i < int_bound.sink.x + n_body; i++)
+	for (unsigned int i = int_bound.sink.x; i < int_bound.sink.x + n_body; i++)
 	{
 		var4_t dVec = {0.0, 0.0, 0.0, 0.0};
-		for (int j = i + 1; j < int_bound.sink.x + n_body; j++) 
+		for (unsigned int j = i + 1; j < int_bound.sink.x + n_body; j++) 
 		{
 			// r_ij 3 FLOP
 			dVec.x = r[j].x - r[i].x;
@@ -1156,10 +1156,10 @@ void cpu_calc_grav_accel_naive_sym_advanced(interaction_bound int_bound, const v
 
 	if (n_sink < n_source)
 	{
-		for (int i = int_bound.sink.x; i < int_bound.sink.y; i++)
+		for (unsigned int i = int_bound.sink.x; i < int_bound.sink.y; i++)
 		{
 			var4_t dVec = {0.0, 0.0, 0.0, 0.0};
-			for (int j = int_bound.source.x + n_body; j < int_bound.source.y; j++) 
+			for (unsigned int j = int_bound.source.x + n_body; j < int_bound.source.y; j++) 
 			{
 				// r_ij 3 FLOP
 				dVec.x = r[j].x - r[i].x;
@@ -1183,10 +1183,10 @@ void cpu_calc_grav_accel_naive_sym_advanced(interaction_bound int_bound, const v
 	}
 	else
 	{
-		for (int i = int_bound.sink.x + n_body; i < int_bound.sink.y; i++)
+		for (unsigned int i = int_bound.sink.x + n_body; i < int_bound.sink.y; i++)
 		{
 			var4_t dVec = {0.0, 0.0, 0.0, 0.0};
-			for (int j = int_bound.source.x; j < int_bound.source.y; j++) 
+			for (unsigned int j = int_bound.source.x; j < int_bound.source.y; j++) 
 			{
 				// r_ij 3 FLOP
 				dVec.x = r[j].x - r[i].x;
@@ -2045,8 +2045,8 @@ void compare_results(option_t& opt, bool verbose)
 		copy_vector_to_device(d_m, h_m, n_total*sizeof(var_t));
 	}
 
-	uint2_t sink   = {0, opt.n0};
-	uint2_t source = {0, opt.n1};
+	uint2_t sink   = {0, (unsigned int)opt.n0};
+	uint2_t source = {0, (unsigned int)opt.n1};
 	interaction_bound int_bound(sink, source);
 
 	cout << "tolerance level   = " << scientific << opt.tol << endl;
@@ -2134,8 +2134,8 @@ void benchmark(int id_dev, int n0, int n1, int dn, int n_iter, ofstream& o_resul
 			copy_vector_to_device(d_x, h_x, n_total*sizeof(var4_t));
 			copy_vector_to_device(d_m, h_m, n_total*sizeof(var_t));
 
-			uint2_t sink = {0, n_snk};
-			uint2_t source = {0, n_src};
+			uint2_t sink   = {0, (unsigned int)n_snk};
+			uint2_t source = {0, (unsigned int)n_src};
 			interaction_bound int_bound(sink, source);
 
 			cout << "(n_sink * n_source = " << setw(6) << n_snk << " * " << setw(6) << n_src << " = " << setw(12) << (uint32_t)n_snk*(uint32_t)n_src << ")---------------------------------------------------------------" << endl;
@@ -2219,8 +2219,8 @@ void benchmark_GPU(option& opt, ofstream& o_result, ofstream& o_summary)
 			copy_vector_to_device(d_x, h_x, n_total*sizeof(var4_t));
 			copy_vector_to_device(d_m, h_m, n_total*sizeof(var_t));
 
-			uint2_t sink = {0, n_snk};
-			uint2_t source = {0, n_src};
+			uint2_t sink   = {0, (unsigned int)n_snk};
+			uint2_t source = {0, (unsigned int)n_src};
 			interaction_bound int_bound(sink, source);
 
 			cout << "(n_sink * n_source = " << setw(6) << n_snk << " * " << setw(6) << n_src << " = " << setw(12) << (uint32_t)n_snk*(uint32_t)n_src << ")-------------------------------------------------------------------" << endl;
@@ -2283,8 +2283,8 @@ void benchmark_CPU(option& opt, ofstream& o_result, ofstream& o_summary)
 
 			populate(n_total, h_x, h_m);
 
-			uint2_t sink = {0, n_snk};
-			uint2_t source = {0, n_src};
+			uint2_t sink   = {0, (unsigned int)n_snk};
+			uint2_t source = {0, (unsigned int)n_src};
 			interaction_bound int_bound(sink, source);
 
 			cout << "(n_sink * n_source = " << setw(6) << n_snk << " * " << setw(6) << n_src << " = " << setw(12) << (uint32_t)n_snk*(uint32_t)n_src << ")-------------------------------------------------------------------" << endl;
